requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
langchain-community
chromadb
pypdf